+++ .env
@@ -15,9 +15,9 @@ AGENT_TEAM5_QUANTITY=0
 AGENT_TEAM6_QUANTITY=0
 AGENT_TEAM4_QUANTITY=0
 AGENT_TEAM3_QUANTITY=0
-AGENT_TEAM3NEUT_QUANTITY=2
-AGENT_TEAM3PAS_QUANTITY=2
-AGENT_TEAM3AGR_QUANTITY=2
+AGENT_TEAM3NEUT_QUANTITY=100
+AGENT_TEAM3PAS_QUANTITY=100
+AGENT_TEAM3AGR_QUANTITY=100
 SELFISH_PER=25
 COLLECTIVE_PER=50
 SELFLESS_PER=75
