+++ pkg/infra/teams/team3/agent3.go
@@ -7,7 +7,6 @@ import (
 	"infra/game/decision"
 	"infra/logging"
 	"math"
-	"math/rand"
 
 	"github.com/benbjohnson/immutable"
 )
@@ -69,25 +68,6 @@ func (a *AgentThree) UpdateInternalState(baseAgent agent.BaseAgent, _ *commons.I
 	if enable {
 		a.UpdatePersonality(baseAgent)
 	}
-}
-
-func (a *AgentThree) Sanctioning() int {
-	return 50
-}
-
-func (a *AgentThree) PruneAgentList(agentMap map[commons.ID]agent.Agent) map[commons.ID]agent.Agent {
-	// fmt.Println("Agent 3")
-	prunned := make(map[commons.ID]agent.Agent)
-	for id, agent := range agentMap {
-		// Compare to 50 in order to sanction
-		toSanctionOrNot := rand.Intn(100)
-		if toSanctionOrNot > a.Sanctioning() {
-			prunned[id] = agent
-		}
-	}
-	// fmt.Println(len(agentMap))
-	// fmt.Println(len(prunned))
-	return prunned
 }
 
 func (a *AgentThree) UpdatePersonality(baseAgent agent.BaseAgent) {
